Fix for duplicate 'View' file(s)
Can't compile with two files named the same.
