+++ Sources/iOS-Common-Libraries/Extensions/View.swift
@@ -3,6 +3,7 @@
 //  iOS-Common-Libraries
 //
 //  Created by Dinesh Harjani on 9/8/22.
+//  Copyright © 2022 Nordic Semiconductor. All rights reserved.
 //
 
 import SwiftUI
@@ -18,6 +19,8 @@ public typealias FormIniOSListInMacOS = List
 // MARK: - View
 
 public extension View {
+    
+    // MARK: frame
     
     @inlinable func frame(size: CGSize, alignment: Alignment = .center) -> some View {
         return frame(width: size.width, height: size.height, alignment: alignment)
@@ -65,6 +68,31 @@ public extension View {
         #if os(iOS)
             .autocapitalization(.none)
         #endif
+    }
+}
+
+// MARK: - doOnce()
+
+public extension View {
+    
+    func doOnce(_ action: @escaping () -> Void) -> some View {
+        modifier(OnceOnly(action: action))
+    }
+}
+
+private struct OnceOnly: ViewModifier {
+    
+    let action: () -> Void
+    
+    @State private var doneAlready = false
+    
+    func body(content: Content) -> some View {
+        // And then, track it here
+        content.onAppear {
+            guard !doneAlready else { return }
+            doneAlready = true
+            action()
+        }
     }
 }
